Correct serial version

+++ threads/hull.h
@@ -4,10 +4,10 @@
 struct Line;
 
 struct Point {
-    long long x, y;
+    double x, y;
 
     Point(): x(0), y(0) {}
-    Point(long long x, long long y) : x(x), y(y) {}
+    Point(double x, double y) : x(x), y(y) {}
     Point& operator=(Point const &other) = default;
 
     bool operator==(Point const& other) const;
@@ -17,6 +17,6 @@ struct Point {
 };
 
 struct Line {
-    long long a, b, c; // a*x + b*y + c
+    double a, b, c; // a*x + b*y + c
     Line (Point const& p1, Point const& p2);
 };
